requirements.txt
telethon==1.42.0
mistralai==1.10.0
python-dotenv==1.2.1
aiohttp==3.13.2